+++ src/components/desktop/Flight-form.vue
@@ -393,8 +393,8 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-@import "src/assets/variables.scss";
-@import "src/assets/font.scss";
+@import "../../assets/variables";
+@import "../../assets/font";
 
 .hero {
   &-content {
